Fixed cell drawing bug from last revision.
Was doing an Int divide by 2 instead of CGFloat for the image size.

+++ Sudoku/SudokuPuzzle/Drawer.swift
@@ -133,7 +133,7 @@ extension SudokuPuzzle {
             draw( cell: cell, puzzle: puzzle, context: context )
             return NSImage(
                 cgImage: context.makeImage()!,
-                size: NSSize( width: width / 2, height: height / 2 )
+                size: NSSize( width: CGFloat( width ) / 2, height: CGFloat( height ) / 2 )
             )
         }
